:hammer: Refactor(pages): reorganize pages to subfolders to contain future redux models
